move vm pod logic to vm.go
